Added hat command and debug utils
The hat command has been added, however if the players inventory is full
when they remove their hat, it will just delete the item. I will add a
fix for this later.

Some debug stuff has been added to help try to figure out what breaks
when it breaks. Not every class has it, but a couple large ones like
DatabaseUtil have it.